Change the documentation: IMAP port: 993 didn't work, but works on the default IMAP port: 143. Adding more useful information.
The correct IMAP port: 993 didn't work, I've already checked if the port is open or not with netcat and indeed it is. Although it works with the default port (143). I suggest changing the README.md documentation on (on line 104) to the correct ports. Alternatively, you can set up the email server to set the IMAP port to 993. On Another note, Some Application doesn't default the connection security to STARTTLS.

On K-9 Mail (Android) I had to tinker for a while before I ended up with the correct information. In the case of Thunderbird, it does automatically set the ports, etc. It's better to have this on the documentation for easy user setup, just to make sure.