Add spamassassin start and enable at boot.
There appears to be an issue when using Debian 11 where spamassassin won't register to start at boot, and even in a post-install rebooted state, will not be running.  Perhaps I am incorrect, but I think this is a useful addition that makes sure this is all set up to operate as you intended.  I am not much of a systemd guy, but I think this is the correct way to do this.  It appears to work on a server I used this script on.  Please reject if superfluous or unnecessary.

Thanks for your work automating the "really fun" part of all this!